requirements.txt
psycopg2-binary
PyGithub
python-dotenv
gql[all]